Feature: SQLDelight DB

+++ app/src/main/kotlin/me/nemiron/khinkalyator/features/people/domain/DeletePersonUseCase.kt
@@ -1,16 +1,20 @@
 package me.nemiron.khinkalyator.features.people.domain
 
+import me.nemiron.khinkalyator.common_domain.MeetsStorage
+import me.nemiron.khinkalyator.common_domain.PeopleStorage
+import me.nemiron.khinkalyator.common_domain.exception.PersonInActiveMeetException
+import me.nemiron.khinkalyator.common_domain.model.PersonId
+
 class DeletePersonUseCase(
-    private val peopleStorage: PeopleStorage
+    private val peopleStorage: PeopleStorage,
+    private val meetsStorage: MeetsStorage
 ) {
     suspend operator fun invoke(personId: PersonId) {
-        /*
-        * TODO: check that person doesn't attempt in any Meet
-        *  if true – just delete it from storage
-        *  if false – replace it with Person.DELETED with previous id and name
-        *  resource(R.string.person_deleted, deletedCount+1),
-        *  deletedCount depends on count of already deleted person
-        * */
-        peopleStorage.deletePerson(personId)
+        val isPersonAttemptsInAnyMeet = meetsStorage.isPersonAttemptsInAnyMeet(personId)
+        if (isPersonAttemptsInAnyMeet) {
+            throw PersonInActiveMeetException
+        } else {
+            peopleStorage.deletePerson(personId)
+        }
     }
 }
